djangoapp/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-jazzmin==3.0.1
django-unfold==0.66.0
djangorestframework==3.14.0
hashids==1.3.1
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
